+++ src/main/java/io/spring/api/article/validation/DuplicatedArticleValidator.java
@@ -1,4 +1,4 @@
-package io.spring.application.article;
+package io.spring.api.article.validation;
 
 import io.spring.core.article.Article;
 import io.spring.core.article.ArticleRepository;
